resource/bot_data/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
configparser==5.2.0
DateTime==4.7
emoji==1.7.0
Flask==2.1.1
frozenlist==1.3.0
ftfy==6.1.1
future==0.18.2
gevent==21.12.0
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
line-bot-sdk==2.2.1
MarkupSafe==2.1.1
multidict==6.0.2
python-Levenshtein==0.12.2
pytz==2022.6
requests==2.27.1
thefuzz==0.19.0
urllib3==1.26.9
wcwidth==0.2.5
Werkzeug==2.1.1
yarl==1.7.2
zipp==3.8.0
zope.event==4.5.0
zope.interface==5.4.0
python-dotenv==1.0.0
